refactor: change the store api prefix to stores
