+++ app/api/stores/exceptions.py
