+++ app/api/stores/service.py
